+++ packages/react-components/react-overflow/src/stories/Overflow/MinimumVisible.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { makeStyles, shorthands } from '@griffel/react';
-import { Overflow } from '../components/Overflow';
+import { Overflow } from '@fluentui/react-overflow';
 import { OverflowMenu, TestOverflowItem } from './utils.stories';
 
 const useStyles = makeStyles({
@@ -12,13 +12,13 @@ const useStyles = makeStyles({
   },
 });
 
-export const DomOrder = () => {
+export const MinimumVisible = () => {
   const styles = useStyles();
 
   const itemIds = new Array(8).fill(0).map((_, i) => i.toString());
 
   return (
-    <Overflow>
+    <Overflow minimumVisible={4}>
       <div className={styles.container}>
         {itemIds.map(i => (
           <TestOverflowItem key={i} id={i} />
