+++ packages/react-components/react-overflow/src/stories/Overflow/ReverseDomOrder.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { makeStyles, shorthands } from '@griffel/react';
-import { Overflow } from '../components/Overflow';
+import { Overflow } from '@fluentui/react-overflow';
 import { OverflowMenu, TestOverflowItem } from './utils.stories';
 
 const useStyles = makeStyles({
@@ -12,18 +12,18 @@ const useStyles = makeStyles({
   },
 });
 
-export const CustomPriorities = () => {
+export const ReverseDomOrder = () => {
   const styles = useStyles();
 
-  const priorities = [2, 3, 6, 1, 4, 5, 0, 7];
+  const itemIds = new Array(8).fill(0).map((_, i) => i.toString());
 
   return (
-    <Overflow>
+    <Overflow overflowDirection="start">
       <div className={styles.container}>
-        {priorities.map(i => (
-          <TestOverflowItem key={i} id={i.toString()} priority={i} />
+        {itemIds.map(i => (
+          <TestOverflowItem key={i} id={i} />
         ))}
-        <OverflowMenu itemIds={priorities.map(x => x.toString())} />
+        <OverflowMenu itemIds={itemIds} />
       </div>
     </Overflow>
   );
