+++ packages/react-components/react-overflow/src/stories/Overflow/utils.stories.tsx
@@ -10,11 +10,13 @@ import {
   MenuItemProps,
   makeStyles,
 } from '@fluentui/react-components';
-import type { OverflowItemProps } from '../components/OverflowItem/OverflowItem.types';
-import { OverflowItem } from '../components/OverflowItem/OverflowItem';
-import { useOverflowMenu } from '../useOverflowMenu';
-import { useIsOverflowItemVisible } from '../useIsOverflowItemVisible';
-import { useIsOverflowGroupVisible } from '../useIsOverflowGroupVisible';
+import type { OverflowItemProps } from '@fluentui/react-overflow';
+import {
+  OverflowItem,
+  useIsOverflowGroupVisible,
+  useIsOverflowItemVisible,
+  useOverflowMenu,
+} from '@fluentui/react-overflow';
 
 const useStyles = makeStyles({
   overflowItem: {
